introduced conditional compilation to remove isolate and ui from the framework on web
